Integrate Phase1 with Phase2

+++ Phase2/headers.h
@@ -59,7 +59,7 @@ struct processStats
     int startTime;
     int waitingTime;
     int TA;
-    int WTA;
+    float WTA;
 };
 
 struct msgbuff
